MINOR: Code cleanup of 'clients' module (#5427)
Cleanup involves:
* Refactoring to use Java 8 constructs (lambdas,
diamond for `empty` collection methods) and library
methods (`computeIfAbsent`)
* Simplifying code (including unnecessarily complex
`equals` and `hashCode` implementations)
* Removing redundant code
* Fixing typos

Reviewers: Ryanne Dolan, Ismael Juma <ismael@juma.me.uk>

+++ clients/src/test/java/org/apache/kafka/common/serialization/SerializationTest.java
@@ -38,7 +38,7 @@ public class SerializationTest {
         {
             put(String.class, Arrays.asList("my string"));
             put(Short.class, Arrays.asList((short) 32767, (short) -32768));
-            put(Integer.class, Arrays.asList((int) 423412424, (int) -41243432));
+            put(Integer.class, Arrays.asList(423412424, -41243432));
             put(Long.class, Arrays.asList(922337203685477580L, -922337203685477581L));
             put(Float.class, Arrays.asList(5678567.12312f, -5678567.12341f));
             put(Double.class, Arrays.asList(5678567.12312d, -5678567.12341d));
